Add simple router and migrate pages to MVC
