+++ controllers/ArticleController.php
@@ -1,0 +1,54 @@
+<?php
+class ArticleController {
+    private $api;
+    public function __construct($api) {
+        $this->api = $api;
+    }
+    public function show($id) {
+        if (!isset($id) || !is_numeric($id)) {
+            header('Location: articles.php');
+            exit();
+        }
+        $article_id = (int)$id;
+        try {
+            $article = $this->api->request('/articles/' . $article_id);
+            $comments = $this->api->request('/articles/' . $article_id . '/comments');
+            $this->api->request('/articles/' . $article_id . '/view', 'POST');
+            $page_title = $article['title'];
+            $page_description = substr(strip_tags($article['content']), 0, 160);
+        } catch (Exception $e) {
+            $_SESSION['flash_message'] = "Erreur: " . $e->getMessage();
+            $_SESSION['flash_type'] = "error";
+            header('Location: articles.php');
+            exit();
+        }
+
+        $comment_error = '';
+        if ($_SERVER['REQUEST_METHOD'] === 'POST' && $this->api->isLoggedIn()) {
+            $content = isset($_POST['content']) ? $_POST['content'] : '';
+            $parent_id = isset($_POST['parent_id']) ? (int)$_POST['parent_id'] : null;
+            if (empty($content)) {
+                $comment_error = "Le contenu du commentaire ne peut pas être vide.";
+            } else {
+                try {
+                    $data = [
+                        'post_id' => $article_id,
+                        'content' => $content
+                    ];
+                    if ($parent_id) {
+                        $data['parent_id'] = $parent_id;
+                    }
+                    $result = $this->api->request('/articles/' . $article_id . '/comments', 'POST', $data, true);
+                    header('Location: article.php?id=' . $article_id . '&comment=success#comment-' . $result['id']);
+                    exit();
+                } catch (Exception $e) {
+                    $comment_error = "Erreur lors de l'envoi de votre commentaire: " . $e->getMessage();
+                }
+            }
+        }
+
+        require __DIR__ . '/../views/templates/header.php';
+        require __DIR__ . '/../views/article.php';
+        require __DIR__ . '/../views/templates/footer.php';
+    }
+}
