+++ edit-article.php
@@ -10,7 +10,7 @@ if (!$api->isLoggedIn()) {
     $_SESSION['redirect_after_login'] = 'my-content.php';
     $_SESSION['flash_message'] = "Vous devez être connecté pour accéder à cette page.";
     $_SESSION['flash_type'] = "error";
-    header('Location: login.php');
+    header('Location: index.php?route=login');
     exit();
 }
